requirements.txt
gradio==4.44.1
gradio-client==1.3.0
xgboost==2.0.3
pandas==2.0.3
numpy==1.25.2
scikit-learn==1.3.2
plotly==5.18.0
google-cloud-storage==2.10.0
requests==2.31.0
fastapi==0.111.0 # Added for FastAPI framework
uvicorn[standard]==0.30.1 # Added as ASGI server for FastAPI
matplotlib==3.8.4 # Added for plotting in the evaluation framework
seaborn==0.13.2 # Added for visualization in the evaluation framework
statsmodels==0.14.1 # Added for LOWESS smoothing in evaluation visualization
pytest==7.4.0
httpx==0.25.0
pytest-asyncio==0.21.0
selenium==4.22.0
webdriver-manager==4.0.1
python-dotenv==1.0.0 # Added for environment variable loading
pyyaml==6.0.1 # Added for YAML configuration
# Pinned versions based on common compatibility, adjust as needed.
